Compute the default width of the gallery after the element is attached to the DOM
In combination with display: block (that implies width 100%) once the element is attached I can get the proper width for the gallery (this needs to be provided for the flicker algorithm). There is still a method that is called if the imput has changed. This method re-renders the gallery if needed.